+++ library-modules/apis/example-tasks/publications/task-publications.ts
@@ -1,5 +1,5 @@
 import { Meteor } from 'meteor/meteor';
-import { TasksCollection } from "../task-collections";
+import { TasksCollection } from "../task-collection";
 import { PublicationIdentifier } from "/library-modules/apis/core/publication-identifier";
 
 Meteor.publish(PublicationIdentifier.TASK, () => {
