+++ library-modules/apis/example-tasks/task-api.ts
@@ -1,5 +1,5 @@
 import { ApiTask } from "./models/api-task";
-import { TasksCollection } from "./task-collections";
+import { TasksCollection } from "./task-collection";
 
 export function getAllApiTasks(): ApiTask[] {
   return TasksCollection.find({}).fetch()
